requirements.txt
certifi==2021.5.30
click==8.0.1
cloudinary==1.26.0
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
